Added try/catch and ALLOWD_ID verification

+++ app/app.py
@@ -20,37 +20,47 @@ bot = TeleBot(BOT_TOKEN)
 
 @bot.message_handler(func=lambda message: True)
 def echo_all(message):
-    result = commandHandler.execute_command(message.text)
-    if message.text.startswith("/d") and result.lower().endswith(".png"):
-        photo = open(str(result),'rb')
-        bot.send_photo(message.chat.id,photo)
-        os.remove(str(result))
-    else:
-        bot.reply_to(message,result, parse_mode='Markdown')
-
+    try:
+        if message.chat.id in ALLOWED_IDS:
+            result = commandHandler.execute_command(message.text)
+            if message.text.startswith("/d") and result.lower().endswith(".png"):
+                photo = open(str(result),'rb')
+                bot.send_photo(message.chat.id,photo)
+                os.remove(str(result))
+            else:
+                bot.reply_to(message,result, parse_mode='Markdown')
+    except Exception as e:
+        logger.error(str(e))
 
 # Handles Voice messages
 @bot.message_handler(content_types=['voice'])
 def handle_docs_audio(message):
-    file_info = bot.get_file(message.voice.file_id)
-    file_path = audio_dir / f'{message.voice.file_unique_id}.ogg'
-    downloaded_file = bot.download_file(file_info.file_path)
-    with open(file_path, 'wb') as new_file:
-        new_file.write(downloaded_file)
-    result = commandHandler.transcript(str(file_path))
-    bot.reply_to(message,result)
+    try:
+        if message.chat.id in ALLOWED_IDS:
+            file_info = bot.get_file(message.voice.file_id)
+            file_path = audio_dir / f'{message.voice.file_unique_id}.ogg'
+            downloaded_file = bot.download_file(file_info.file_path)
+            with open(file_path, 'wb') as new_file:
+                new_file.write(downloaded_file)
+            result = commandHandler.transcript(str(file_path))
+            bot.reply_to(message,result)
+    except Exception as e:
+        logger.error(str(e))
 
 # Handles Audio messages
 @bot.message_handler(content_types=['audio'])
 def handle_docs_audio(message):
-    file_info = bot.get_file(message.audio.file_id)
-    file_path = audio_dir / message.audio.file_name
-    downloaded_file = bot.download_file(file_info.file_path)
-    with open(file_path, 'wb') as new_file:
-        new_file.write(downloaded_file)
-    result = commandHandler.transcript(str(file_path))
-    bot.reply_to(message,result)
-
+    try:
+        if message.chat.id in ALLOWED_IDS:
+            file_info = bot.get_file(message.audio.file_id)
+            file_path = audio_dir / message.audio.file_name
+            downloaded_file = bot.download_file(file_info.file_path)
+            with open(file_path, 'wb') as new_file:
+                new_file.write(downloaded_file)
+            result = commandHandler.transcript(str(file_path))
+            bot.reply_to(message,result)
+    except Exception as e:
+        logger.error(str(e))
